# Conflicts:
#	src/Conduit/Features/Profiles/ProfileReader.cs
#	src/Conduit/Features/Users/Edit.cs
#	tests/Conduit.IntegrationTests/packages.lock.json

+++ src/Conduit/Features/Articles/Delete.cs
@@ -18,19 +18,21 @@ public class Delete
         public CommandValidator() => RuleFor(x => x.Slug).NotNull().NotEmpty();
     }
 
-    public class QueryHandler(ConduitContext context) : IRequestHandler<Command>
+    public class QueryHandler(ConduitContext context, ICurrentUserAccessor currentUserAccessor)
+        : IRequestHandler<Command>
     {
         public async Task Handle(Command message, CancellationToken cancellationToken)
         {
             var article =
-                await context.Articles.FirstOrDefaultAsync(
-                    x => x.Slug == message.Slug,
-                    cancellationToken
-                )
-                ?? throw new RestException(
-                    HttpStatusCode.NotFound,
-                    new { Article = Constants.NOT_FOUND }
-                );
+                await context
+                    .Articles.Include(x => x.Author)
+                    .FirstOrDefaultAsync(x => x.Slug == message.Slug, cancellationToken)
+                ?? throw new RestException(HttpStatusCode.NotFound, "article", Constants.NOT_FOUND);
+
+            if (article.Author?.Username != currentUserAccessor.GetCurrentUsername())
+            {
+                throw new RestException(HttpStatusCode.Forbidden, "article", Constants.FORBIDDEN);
+            }
 
             context.Articles.Remove(article);
             await context.SaveChangesAsync(cancellationToken);
