+++ src/Conduit/Features/Comments/Delete.cs
@@ -19,25 +19,26 @@ public class Delete
         public CommandValidator() => RuleFor(x => x.Slug).NotNull().NotEmpty();
     }
 
-    public class QueryHandler(ConduitContext context) : IRequestHandler<Command>
+    public class QueryHandler(ConduitContext context, ICurrentUserAccessor currentUserAccessor)
+        : IRequestHandler<Command>
     {
         public async Task Handle(Command message, CancellationToken cancellationToken)
         {
             var article =
                 await context
                     .Articles.Include(x => x.Comments)
+                    .ThenInclude(x => x.Author)
                     .FirstOrDefaultAsync(x => x.Slug == message.Slug, cancellationToken)
-                ?? throw new RestException(
-                    HttpStatusCode.NotFound,
-                    new { Article = Constants.NOT_FOUND }
-                );
+                ?? throw new RestException(HttpStatusCode.NotFound, "article", Constants.NOT_FOUND);
 
             var comment =
                 article.Comments.FirstOrDefault(x => x.CommentId == message.Id)
-                ?? throw new RestException(
-                    HttpStatusCode.NotFound,
-                    new { Comment = Constants.NOT_FOUND }
-                );
+                ?? throw new RestException(HttpStatusCode.NotFound, "comment", Constants.NOT_FOUND);
+
+            if (comment.Author?.Username != currentUserAccessor.GetCurrentUsername())
+            {
+                throw new RestException(HttpStatusCode.Forbidden, "comment", Constants.FORBIDDEN);
+            }
 
             context.Comments.Remove(comment);
             await context.SaveChangesAsync(cancellationToken);
